Fix GDAL process stop and launch the webbrowser 1 seconde later to avoid launch without app server

+++ vt_utils_parameters.py
@@ -64,5 +64,5 @@ class Parameters:
         self.all_vectors = None
         self.zommLevel = None
         self.tileSize = None
-        if self.GDALprocess.is_alive():
+        if self.GDALprocess and self.GDALprocess.is_alive():
             self.GDALprocess.terminate()
